Remove JPEG instance of Datapath
Having it has two main cons: (1) Everytime the .drawio file updates,
it should get updated, (2) it is binary and hard to be tracked by
Git.

Also, it is pointless to include and update it in the repository.
Yes, real-time viewing of the datapath is possible using it, but
it should not be that hard to open a brower, type app.diagrams.net
and upload the .drawio file to view it. Plus, if there is a need to
include the datapath image somewhere (e.g. in the documentation or
the report), it could be exported as JPEG (or even PNG, which is IMO
better) at that specific time.